refactor: BetAmount 검증 로직 InputView로 이동 및 record로 변환
